+++ inst/include/common.hpp
@@ -1,10 +1,9 @@
-// common headers: Rcpp (for messages / interrupts) and the linear algebra layer
+// common headers: R's C API glue (messages / interrupts) and linear algebra
 
 #ifndef RWIG_COMMON_H
 #define RWIG_COMMON_H
 
-#include <Rcpp.h>
-
 #include "linalg.hpp"
+#include "r_glue.hpp"
 
 #endif // RWIG_COMMON_H