+++ src/barycenter.cpp
@@ -1,35 +1,31 @@
 
-// this is the file defining the functions exporting to R side
-// barycenter algos
+// .Call entry points for the barycenter algorithms
 
 #include "barycenter_impl.hpp"
-#include "rcpp_glue.hpp"
 
-static Rcpp::List barycenter_result(const Barycenter &bc, bool withgrad,
-                                    const char *Uname, const char *Vname) {
+static SEXP barycenter_result(const Barycenter &bc, bool withgrad,
+                              const char *Uname, const char *Vname) {
+  rr::ListBuilder out;
+  out.add("b", rr::to_R(bc.b));
   if (withgrad) {
-    return Rcpp::List::create(
-        Rcpp::Named("b") = la::to_R(bc.b), Rcpp::Named("grad_A") = la::to_R(bc.grad_A),
-        Rcpp::Named("grad_w") = la::to_R(bc.grad_w), Rcpp::Named("loss") = bc.loss,
-        Rcpp::Named(Uname) = la::to_R(bc.U), Rcpp::Named(Vname) = la::to_R(bc.V),
-        Rcpp::Named("iter") = bc.iter, Rcpp::Named("err") = bc.err,
-        Rcpp::Named("return_status") = bc.return_code);
-  } else {
-    return Rcpp::List::create(
-        Rcpp::Named("b") = la::to_R(bc.b), Rcpp::Named(Uname) = la::to_R(bc.U),
-        Rcpp::Named(Vname) = la::to_R(bc.V), Rcpp::Named("iter") = bc.iter,
-        Rcpp::Named("err") = bc.err,
-        Rcpp::Named("return_status") = bc.return_code);
+    out.add("grad_A", rr::to_R(bc.grad_A))
+        .add("grad_w", rr::to_R(bc.grad_w))
+        .add("loss", bc.loss);
   }
+  out.add(Uname, rr::to_R(bc.U))
+      .add(Vname, rr::to_R(bc.V))
+      .add("iter", bc.iter)
+      .add("err", bc.err)
+      .add("return_status", bc.return_code);
+  return out.build();
 }
 
-Rcpp::List barycenter_parallel_cpu(const SEXP &A, const SEXP &C, const SEXP &w,
-                                   double reg, const SEXP &b_ext,
-                                   bool withgrad = false, int maxiter = 1000,
-                                   double zerotol = 1e-6, int verbose = 0) {
-  la::Mat A_ = la::mat_from_R(A);
-  la::Mat C_ = la::mat_from_R(C);
-  la::Vec w_ = la::vec_from_R(w);
+static SEXP barycenter_parallel_cpu(SEXP A, SEXP C, SEXP w, double reg,
+                                    SEXP b_ext, bool withgrad, int maxiter,
+                                    double zerotol, int verbose) {
+  la::Mat A_ = rr::mat_from_R(A);
+  la::Mat C_ = rr::mat_from_R(C);
+  la::Vec w_ = rr::vec_from_R(w);
 
   // init the class
   Barycenter bc((int)A_.ncol(), withgrad, maxiter, zerotol, verbose);
@@ -39,7 +35,7 @@ Rcpp::List barycenter_parallel_cpu(const SEXP &A, const SEXP &C, const SEXP &w,
   bc.update_A(A_);
   bc.update_w(w_);
   if (withgrad) {
-    bc.update_b_ext(la::vec_from_R(b_ext));
+    bc.update_b_ext(rr::vec_from_R(b_ext));
   }
 
   // start the computation
@@ -53,51 +49,32 @@ Rcpp::List barycenter_parallel_cpu(const SEXP &A, const SEXP &C, const SEXP &w,
 
 #include "cuda_interface.cuh"
 
-Rcpp::List barycenter_parallel_cuda(const SEXP &A, const SEXP &C, const SEXP &w,
-                                    double reg, const SEXP &b_ext,
-                                    bool withgrad = false, int maxiter = 1000,
-                                    double zerotol = 1e-6, int verbose = 0) {
-  double *A_ptr = REAL(A);
-  double *C_ptr = REAL(C);
-  double *w_ptr = REAL(w);
-  double *b_ext_ptr = REAL(b_ext);
-  int M = Rf_nrows(C);
-  int N = Rf_ncols(C);
-  int S = Rf_ncols(A);
+static SEXP barycenter_parallel_cuda(SEXP A, SEXP C, SEXP w, double reg,
+                                     SEXP b_ext, bool withgrad, int maxiter,
+                                     double zerotol, int verbose) {
+  rr::Protector p;
+  A = rr::as_real(A, p);
+  C = rr::as_real(C, p);
+  w = rr::as_real(w, p);
+  b_ext = rr::as_real(b_ext, p);
+  const int M = Rf_nrows(C);
+  const int N = Rf_ncols(C);
+  const int S = Rf_ncols(A);
 
-  SEXP b_ = PROTECT(Rf_allocVector(REALSXP, N));
-  SEXP U_ = PROTECT(Rf_allocVector(REALSXP, M * S));
-  SEXP V_ = PROTECT(Rf_allocVector(REALSXP, N * S));
-  SEXP grad_A_ = PROTECT(Rf_allocVector(REALSXP, M * S));
-  SEXP grad_w_ = PROTECT(Rf_allocVector(REALSXP, S));
-
-  double *b_ptr = REAL(b_);
-  double *U_ptr = REAL(U_);
-  double *V_ptr = REAL(V_);
-  double *grad_A_ptr = REAL(grad_A_);
-  double *grad_w_ptr = REAL(grad_w_);
+  SEXP b_ = rr::alloc_vector(N, p);
+  SEXP U_ = rr::alloc_matrix(M, S, p);
+  SEXP V_ = rr::alloc_matrix(N, S, p);
+  SEXP grad_A_ = rr::alloc_matrix(M, S, p);
+  SEXP grad_w_ = rr::alloc_vector(S, p);
 
   double loss = 0.;
   int iter = 0;
   double err = 0.;
 
-  cuda_barycenter_parallel(U_ptr, V_ptr, b_ptr, grad_A_ptr, grad_w_ptr, &loss,
-                           &iter, &err, A_ptr, w_ptr, C_ptr, b_ext_ptr, M, N, S,
-                           reg, withgrad, maxiter, zerotol);
-
-  // set dims
-  SEXP dims_MS = PROTECT(Rf_allocVector(INTSXP, 2));
-  INTEGER(dims_MS)[0] = M;
-  INTEGER(dims_MS)[1] = S;
-  SEXP dims_NS = PROTECT(Rf_allocVector(INTSXP, 2));
-  INTEGER(dims_NS)[0] = N;
-  INTEGER(dims_NS)[1] = S;
-
-  Rf_setAttrib(U_, R_DimSymbol, dims_MS);
-  Rf_setAttrib(V_, R_DimSymbol, dims_NS);
-  Rf_setAttrib(grad_A_, R_DimSymbol, dims_MS);
-
-  UNPROTECT(7);
+  cuda_barycenter_parallel(REAL(U_), REAL(V_), REAL(b_), REAL(grad_A_),
+                           REAL(grad_w_), &loss, &iter, &err, REAL(A), REAL(w),
+                           REAL(C), REAL(b_ext), M, N, S, reg, withgrad,
+                           maxiter, zerotol);
 
   int return_code;
   if (err <= zerotol) {
@@ -108,19 +85,12 @@ Rcpp::List barycenter_parallel_cuda(const SEXP &A, const SEXP &C, const SEXP &w,
     return_code = 2;
   }
 
-  if (withgrad) {
-    return Rcpp::List::create(
-        Rcpp::Named("b") = b_, Rcpp::Named("grad_A") = grad_A_,
-        Rcpp::Named("grad_w") = grad_w_, Rcpp::Named("loss") = loss,
-        Rcpp::Named("U") = U_, Rcpp::Named("V") = V_,
-        Rcpp::Named("iter") = iter, Rcpp::Named("err") = err,
-        Rcpp::Named("return_status") = return_code);
-  } else {
-    return Rcpp::List::create(Rcpp::Named("b") = b_, Rcpp::Named("U") = U_,
-                              Rcpp::Named("V") = V_, Rcpp::Named("iter") = iter,
-                              Rcpp::Named("err") = err,
-                              Rcpp::Named("return_status") = return_code);
-  }
+  rr::ListBuilder out;
+  out.add("b", b_);
+  if (withgrad) out.add("grad_A", grad_A_).add("grad_w", grad_w_).add("loss", loss);
+  out.add("U", U_).add("V", V_).add("iter", iter).add("err", err);
+  out.add("return_status", return_code);
+  return out.build();
 }
 
 #endif
@@ -129,52 +99,54 @@ Rcpp::List barycenter_parallel_cuda(const SEXP &A, const SEXP &C, const SEXP &w,
 Interfaces for the R side
 */
 
-// [[Rcpp::export]]
-Rcpp::List barycenter_parallel_cpp(const SEXP &A, const SEXP &C, const SEXP &w,
-                                   double reg, const SEXP &b_ext,
-                                   bool withgrad = false, bool usecuda = true,
-                                   int maxiter = 1000, double zerotol = 1e-6,
-                                   int verbose = 0) {
-  Rcpp::List res;
-
+extern "C" SEXP rwig_barycenter_parallel_cpp(SEXP A, SEXP C, SEXP w, SEXP reg,
+                                             SEXP b_ext, SEXP withgrad,
+                                             SEXP usecuda, SEXP maxiter,
+                                             SEXP zerotol, SEXP verbose) {
+  return rr::call_guard([&]() -> SEXP {
+    const double reg_ = rr::as_double(reg);
+    const bool withgrad_ = rr::as_bool(withgrad);
+    const int maxiter_ = rr::as_int(maxiter);
+    const double zerotol_ = rr::as_double(zerotol);
+    const int verbose_ = rr::as_int(verbose);
 #if defined(HAVE_CUBLAS) && defined(HAVE_CUDA_RUNTIME)
-  if (usecuda) {
-    res = barycenter_parallel_cuda(A, C, w, reg, b_ext, withgrad, maxiter,
-                                   zerotol, verbose);
-  } else {
-    res = barycenter_parallel_cpu(A, C, w, reg, b_ext, withgrad, maxiter,
-                                  zerotol, verbose);
-  }
+    if (rr::as_bool(usecuda)) {
+      return barycenter_parallel_cuda(A, C, w, reg_, b_ext, withgrad_, maxiter_,
+                                      zerotol_, verbose_);
+    }
 #else
-  res = barycenter_parallel_cpu(A, C, w, reg, b_ext, withgrad, maxiter, zerotol,
-                                verbose);
+    (void)usecuda;
 #endif
-  return res;
+    return barycenter_parallel_cpu(A, C, w, reg_, b_ext, withgrad_, maxiter_,
+                                   zerotol_, verbose_);
+  });
 }
 
-// [[Rcpp::export]]
-Rcpp::List barycenter_log_cpp(const SEXP &A, const SEXP &C, const SEXP &w,
-                              double reg, const SEXP &b_ext,
-                              bool withgrad = false, const int &n_threads = 0,
-                              int maxiter = 1000, double zerotol = 1e-6,
-                              int verbose = 0) {
-  la::Mat A_ = la::mat_from_R(A);
-  la::Mat C_ = la::mat_from_R(C);
-  la::Vec w_ = la::vec_from_R(w);
+extern "C" SEXP rwig_barycenter_log_cpp(SEXP A, SEXP C, SEXP w, SEXP reg,
+                                        SEXP b_ext, SEXP withgrad,
+                                        SEXP n_threads, SEXP maxiter,
+                                        SEXP zerotol, SEXP verbose) {
+  return rr::call_guard([&]() -> SEXP {
+    la::Mat A_ = rr::mat_from_R(A);
+    la::Mat C_ = rr::mat_from_R(C);
+    la::Vec w_ = rr::vec_from_R(w);
+    const bool withgrad_ = rr::as_bool(withgrad);
 
-  // init the class
-  Barycenter bc((int)A_.ncol(), withgrad, maxiter, zerotol, verbose);
-  // update/load all the data
-  bc.update_C(C_);
-  bc.update_reg(reg);
-  bc.update_A(A_);
-  bc.update_w(w_);
-  if (withgrad) {
-    bc.update_b_ext(la::vec_from_R(b_ext));
-  }
+    // init the class
+    Barycenter bc((int)A_.ncol(), withgrad_, rr::as_int(maxiter),
+                  rr::as_double(zerotol), rr::as_int(verbose));
+    // update/load all the data
+    bc.update_C(C_);
+    bc.update_reg(rr::as_double(reg));
+    bc.update_A(A_);
+    bc.update_w(w_);
+    if (withgrad_) {
+      bc.update_b_ext(rr::vec_from_R(b_ext));
+    }
 
-  // start the computation
-  bc.compute_log(n_threads);
+    // start the computation
+    bc.compute_log(rr::as_int(n_threads));
 
-  return barycenter_result(bc, withgrad, "F", "G");
+    return barycenter_result(bc, withgrad_, "F", "G");
+  });
 }