+++ src/barycenter_impl_log.cpp
@@ -112,12 +112,12 @@ void Barycenter::_fwd_log(ThreadPool &pool) {
 
   // logging for forward pass
   if (_verbose != 0) {
-    Rcpp::message(Rf_mkString("Forward pass:"));
+    rr::message(("Forward pass:"));
   }
   _minrow(pool, this->U, this->V, _KV); // update _KV
 
   while ((this->iter < _maxiter) && (this->err >= _zerotol)) {
-    Rcpp::checkUserInterrupt();
+    rr::check_interrupt();
     this->iter++;
     if (_verbose != 0) {
       _timer.tic();
@@ -207,7 +207,7 @@ void Barycenter::_bwd_log(ThreadPool &pool) {
 
   // logging for backward pass
   if (_verbose != 0) {
-    Rcpp::message(Rf_mkString("Backward pass:"));
+    rr::message(("Backward pass:"));
   }
 
   for (int l = this->iter; l > 0; --l) {