+++ src/wdl.cpp
@@ -1,17 +1,9 @@
 
-// this is the file defining the functions exporting to R side
-// WDL algos
+// .Call entry point for the Wasserstein Dictionary Learning model
 
-// #include <iostream> // std::cout
-// #include "R_ext/Print.h"    // for REprintf
+#include <stdexcept>
 
-#include "rcpp_glue.hpp"
 #include "wdl_impl.hpp" // header for the WDL definition
-// #include "ctrack.hpp"
-
-// using namespace cpp11;
-// using namespace cpp11::literals; // so we can use ""_nm syntax
-// namespace writable = cpp11::writable;
 
 /////////////////////////////////////////////////////////////////////
 // R interfaces for WDL
@@ -19,34 +11,36 @@
 
 // TODO: add warning for the non-converging Barycenter?
 
-Rcpp::List wdl_cpp_cpu(const SEXP &Y, // document matrix N * M
-                       const SEXP &C, // cost matrix N * N
-                       const double reg, const int S, const int n_threads,
-                       const int batch_size, const int epochs,
-                       int sinkhorn_mode = 1, const int max_iter = 1000,
-                       const double zero_tol = 1e-6, const int optimizer = 2,
-                       const double eta = .001, const double gamma = .01,
-                       const double beta1 = .9, const double beta2 = .999,
-                       const double eps = 1e-8, const bool verbose = false) {
-  la::Mat Y_ = la::mat_from_R(Y);
-  la::Mat C_ = la::mat_from_R(C);
+struct WdlArgs {
+  double reg;
+  int S, n_threads, batch_size, epochs, sinkhorn_mode, max_iter, optimizer;
+  double zero_tol, eta, gamma, beta1, beta2, eps;
+  bool verbose;
+  int seed;
+};
+
+static SEXP wdl_cpp_cpu(SEXP Y, SEXP C, const WdlArgs &a) {
+  la::Mat Y_ = rr::mat_from_R(Y);
+  la::Mat C_ = rr::mat_from_R(C);
 
   // init the WDL class
-  WassersteinDictionaryLearning wdl(
-      batch_size, epochs, n_threads, sinkhorn_mode, max_iter, zero_tol,
-      optimizer, eta, gamma, beta1, beta2, eps, verbose);
+  WassersteinDictionaryLearning wdl(a.batch_size, a.epochs, a.n_threads,
+                                    a.sinkhorn_mode, a.max_iter, a.zero_tol,
+                                    a.optimizer, a.eta, a.gamma, a.beta1,
+                                    a.beta2, a.eps, a.verbose);
 
   //  with data
-  wdl.init_data(Y_, C_, reg, S);
+  wdl.init_data(Y_, C_, a.reg, a.S);
 
-  // start the actual WDL computation
-  wdl.compute();
+  // start the actual WDL computation (random init uses R's RNG)
+  {
+    rr::RNGScope rng;
+    wdl.compute();
+  }
 
-  // ctrack::result_print();
-
-  return Rcpp::List::create(Rcpp::Named("A") = la::to_R(wdl.A),
-                            Rcpp::Named("W") = la::to_R(wdl.W),
-                            Rcpp::Named("Yhat") = la::to_R(wdl.Yhat));
+  rr::ListBuilder out;
+  out.add("A", rr::to_R(wdl.A)).add("W", rr::to_R(wdl.W)).add("Yhat", rr::to_R(wdl.Yhat));
+  return out.build();
 }
 
 // only have the CUDA version when they are detected
@@ -54,48 +48,25 @@ Rcpp::List wdl_cpp_cpu(const SEXP &Y, // document matrix N * M
 
 #include "cuda_interface.cuh"
 
-Rcpp::List wdl_cpp_cuda(const SEXP &Y, // document matrix N * M
-                        const SEXP &C, // cost matrix N * N
-                        const double reg, const int S, const int batch_size,
-                        const int epochs, int sinkhorn_mode = 1,
-                        const int max_iter = 1000, const double zero_tol = 1e-6,
-                        const int optimizer = 2, const double eta = .001,
-                        const double gamma = .01, const double beta1 = .9,
-                        const double beta2 = .999, const double eps = 1e-8,
-                        const bool verbose = false, const int seed = 42) {
-  double *Y_ptr = REAL(Y);
-  double *C_ptr = REAL(C);
-  int N = Rf_nrows(Y);
-  int M = Rf_ncols(Y);
+static SEXP wdl_cpp_cuda(SEXP Y, SEXP C, const WdlArgs &a) {
+  rr::Protector p;
+  Y = rr::as_real(Y, p);
+  C = rr::as_real(C, p);
+  const int N = Rf_nrows(Y);
+  const int M = Rf_ncols(Y);
 
   // allocate output matrices
-  SEXP A_ = PROTECT(Rf_allocVector(REALSXP, N * S));
-  SEXP W_ = PROTECT(Rf_allocVector(REALSXP, S * M));
-  SEXP Yhat_ = PROTECT(Rf_allocVector(REALSXP, N * M));
+  SEXP A_ = rr::alloc_matrix(N, a.S, p);
+  SEXP W_ = rr::alloc_matrix(a.S, M, p);
+  SEXP Yhat_ = rr::alloc_matrix(N, M, p);
 
-  cuda_wdl(REAL(A_), REAL(W_), REAL(Yhat_), Y_ptr, C_ptr, N, M, S, reg,
-           max_iter, zero_tol, batch_size, epochs, optimizer, eta, gamma, beta1,
-           beta2, eps, seed, verbose);
+  cuda_wdl(REAL(A_), REAL(W_), REAL(Yhat_), REAL(Y), REAL(C), N, M, a.S, a.reg,
+           a.max_iter, a.zero_tol, a.batch_size, a.epochs, a.optimizer, a.eta,
+           a.gamma, a.beta1, a.beta2, a.eps, a.seed, a.verbose);
 
-  // set dims
-  SEXP dims_NS = PROTECT(Rf_allocVector(INTSXP, 2));
-  INTEGER(dims_NS)[0] = N;
-  INTEGER(dims_NS)[1] = S;
-  SEXP dims_SM = PROTECT(Rf_allocVector(INTSXP, 2));
-  INTEGER(dims_SM)[0] = S;
-  INTEGER(dims_SM)[1] = M;
-  SEXP dims_NM = PROTECT(Rf_allocVector(INTSXP, 2));
-  INTEGER(dims_NM)[0] = N;
-  INTEGER(dims_NM)[1] = M;
-
-  Rf_setAttrib(A_, R_DimSymbol, dims_NS);
-  Rf_setAttrib(W_, R_DimSymbol, dims_SM);
-  Rf_setAttrib(Yhat_, R_DimSymbol, dims_NM);
-
-  UNPROTECT(6);
-
-  return Rcpp::List::create(Rcpp::Named("A") = A_, Rcpp::Named("W") = W_,
-                            Rcpp::Named("Yhat") = Yhat_);
+  rr::ListBuilder out;
+  out.add("A", A_).add("W", W_).add("Yhat", Yhat_);
+  return out.build();
 }
 
 #endif
@@ -104,67 +75,61 @@ Rcpp::List wdl_cpp_cuda(const SEXP &Y, // document matrix N * M
 Interfaces for the R side
 */
 
-// [[Rcpp::export]]
-Rcpp::List wdl_cpp(const SEXP &Y, // document matrix N * M
-                   const SEXP &C, // cost matrix N * N
-                   const double reg, const int S, const int n_threads,
-                   const int batch_size, const int epochs,
-                   int sinkhorn_mode = 1, bool usecuda = true,
-                   const int max_iter = 1000, const double zero_tol = 1e-6,
-                   const int optimizer = 2, const double eta = .001,
-                   const double gamma = .01, const double beta1 = .9,
-                   const double beta2 = .999, const double eps = 1e-8,
-                   const bool verbose = false, const int seed = 42) {
+extern "C" SEXP rwig_wdl_cpp(SEXP Y, SEXP C, SEXP reg, SEXP S, SEXP n_threads,
+                             SEXP batch_size, SEXP epochs, SEXP sinkhorn_mode,
+                             SEXP usecuda, SEXP max_iter, SEXP zero_tol,
+                             SEXP optimizer, SEXP eta, SEXP gamma, SEXP beta1,
+                             SEXP beta2, SEXP eps, SEXP verbose, SEXP seed) {
+  return rr::call_guard([&]() -> SEXP {
+    WdlArgs a;
+    a.reg = rr::as_double(reg);
+    a.S = rr::as_int(S);
+    a.n_threads = rr::as_int(n_threads);
+    a.batch_size = rr::as_int(batch_size);
+    a.epochs = rr::as_int(epochs);
+    a.sinkhorn_mode = rr::as_int(sinkhorn_mode);
+    a.max_iter = rr::as_int(max_iter);
+    a.zero_tol = rr::as_double(zero_tol);
+    a.optimizer = rr::as_int(optimizer);
+    a.eta = rr::as_double(eta);
+    a.gamma = rr::as_double(gamma);
+    a.beta1 = rr::as_double(beta1);
+    a.beta2 = rr::as_double(beta2);
+    a.eps = rr::as_double(eps);
+    a.verbose = rr::as_bool(verbose);
+    a.seed = rr::as_int(seed);
 
-  // check sinkhorn mode
-  if ((sinkhorn_mode != 1) && (sinkhorn_mode != 2)) {
-    Rcpp::stop("Sinkhorn mode not supported");
-  }
+    // check sinkhorn mode
+    if ((a.sinkhorn_mode != 1) && (a.sinkhorn_mode != 2)) {
+      throw std::runtime_error("Sinkhorn mode not supported");
+    }
 
-  // check optimizer mode
-  if ((optimizer != 0) && (optimizer != 1) && (optimizer != 2)) {
-    Rcpp::stop("optimizer must be: 0, 1, 2!");
-  }
+    // check optimizer mode
+    if ((a.optimizer != 0) && (a.optimizer != 1) && (a.optimizer != 2)) {
+      throw std::runtime_error("optimizer must be: 0, 1, 2!");
+    }
 
-  // stop if N <= S
-  if (Rf_nrows(Y) <= S) {
-    Rcpp::stop("Number of topics S must be smaller than the vocab size N!");
-  }
-
-  Rcpp::List res;
+    // stop if N <= S
+    if (Rf_nrows(Y) <= a.S) {
+      throw std::runtime_error(
+          "Number of topics S must be smaller than the vocab size N!");
+    }
 
 #if defined(HAVE_CUBLAS) && defined(HAVE_CUDA_RUNTIME)
-  if (usecuda) {
-    if (verbose) {
-      Rcpp::message(Rf_mkString("Running WDL in CUDA mode..."));
-      Rcpp::message(Rf_mkString(
-          "This might take a while depending on the problem size..."));
+    if (rr::as_bool(usecuda)) {
+      if (a.verbose) {
+        rr::message("Running WDL in CUDA mode...");
+        rr::message("This might take a while depending on the problem size...");
+      }
+      return wdl_cpp_cuda(Y, C, a);
     }
-
-    res = wdl_cpp_cuda(Y, C, reg, S, batch_size, epochs, sinkhorn_mode,
-                       max_iter, zero_tol, optimizer, eta, gamma, beta1, beta2,
-                       eps, verbose, seed);
-  } else {
-    if (verbose) {
-      Rcpp::message(Rf_mkString("Running WDL in CPU mode..."));
-      Rcpp::message(Rf_mkString(
-          "This might take a while depending on the problem size..."));
-    }
-    res = wdl_cpp_cpu(Y, C, reg, S, n_threads, batch_size, epochs,
-                      sinkhorn_mode, max_iter, zero_tol, optimizer, eta, gamma,
-                      beta1, beta2, eps, verbose);
-  }
 #else
-  if (verbose) {
-    Rcpp::message(Rf_mkString("Running WDL in CPU mode..."));
-    Rcpp::message(Rf_mkString(
-        "This might take a while depending on the problem size..."));
-  }
-
-  res = wdl_cpp_cpu(Y, C, reg, S, n_threads, batch_size, epochs, sinkhorn_mode,
-                    max_iter, zero_tol, optimizer, eta, gamma, beta1, beta2,
-                    eps, verbose);
+    (void)usecuda;
 #endif
-
-  return res;
+    if (a.verbose) {
+      rr::message("Running WDL in CPU mode...");
+      rr::message("This might take a while depending on the problem size...");
+    }
+    return wdl_cpp_cpu(Y, C, a);
+  });
 }